Move EDH fitness functions into edh subfolder
