+++ source/api/fitness-leaderboard/justgiving/index.js
@@ -1,0 +1,11 @@
+export const fetchFitnessLeaderboard = (params = required()) => {
+  return Promise.reject(
+    new Error('This method is not supported for JustGiving')
+  )
+}
+
+export const deserializeFitnessLeaderboard = (result, index) => {
+  return Promise.reject(
+    new Error('This method is not supported for JustGiving')
+  )
+}
